+++ docs/html/search/all_c.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['_7ecommanddispatcher_25',['~CommandDispatcher',['../classcli_1_1CommandDispatcher.html#a7e680654e4c8814eff5b7dac67ac2111',1,'cli::CommandDispatcher']]]
+  ['setup_34',['setup',['../classace__utils_1_1mode__group_1_1ModeNavigator.html#a8fcc09eaa24f32eed4d3bb9bcef48f0c',1,'ace_utils::mode_group::ModeNavigator']]],
+  ['streamlinereader_35',['StreamLineReader',['../classace__utils_1_1cli_1_1StreamLineReader.html',1,'ace_utils::cli::StreamLineReader'],['../classace__utils_1_1cli_1_1StreamLineReader.html#ad16d26a6e4286dc4fc2015fe9c68fee1',1,'ace_utils::cli::StreamLineReader::StreamLineReader()']]]
 ];