+++ docs/html/search/classes_0.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['commanddispatcher_26',['CommandDispatcher',['../classcli_1_1CommandDispatcher.html',1,'cli']]],
-  ['commandhandler_27',['CommandHandler',['../classcli_1_1CommandHandler.html',1,'cli']]],
-  ['commandmanager_28',['CommandManager',['../classcli_1_1CommandManager.html',1,'cli']]],
-  ['crceeprom_29',['CrcEeprom',['../classcrc__eeprom_1_1CrcEeprom.html',1,'crc_eeprom']]]
+  ['commanddispatcher_39',['CommandDispatcher',['../classace__utils_1_1cli_1_1CommandDispatcher.html',1,'ace_utils::cli']]],
+  ['commandhandler_40',['CommandHandler',['../classace__utils_1_1cli_1_1CommandHandler.html',1,'ace_utils::cli']]],
+  ['commandmanager_41',['CommandManager',['../classace__utils_1_1cli_1_1CommandManager.html',1,'ace_utils::cli']]],
+  ['crceeprom_42',['CrcEeprom',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html',1,'ace_utils::crc_eeprom']]]
 ];