+++ docs/html/search/functions_2.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['findandruncommand_36',['findAndRunCommand',['../classcli_1_1CommandDispatcher.html#a75d938737d913bacf298424ead78ece8',1,'cli::CommandDispatcher']]],
-  ['findcommand_37',['findCommand',['../classcli_1_1CommandDispatcher.html#abb91486d151eebed9c38c8415cdb05d8',1,'cli::CommandDispatcher']]]
+  ['findandruncommand_54',['findAndRunCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a97631715e21c262544a228377688d593',1,'ace_utils::cli::CommandDispatcher']]],
+  ['findcommand_55',['findCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ad59ab4795f6834ccf252dce8b0682c52',1,'ace_utils::cli::CommandDispatcher']]]
 ];