+++ docs/html/search/functions_4.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['helpall_41',['helpAll',['../classcli_1_1CommandDispatcher.html#aa176113e9282337b422303b8e592d74c',1,'cli::CommandDispatcher']]],
-  ['helpcommandhandler_42',['helpCommandHandler',['../classcli_1_1CommandDispatcher.html#a38a6973e8a181c48db1b988c3dbd61df',1,'cli::CommandDispatcher']]],
-  ['helpspecific_43',['helpSpecific',['../classcli_1_1CommandDispatcher.html#a4c3f9282cfad17a75cce145a5566a01b',1,'cli::CommandDispatcher']]]
+  ['helpall_59',['helpAll',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a517055113aecc022baed1fa53742ee7c',1,'ace_utils::cli::CommandDispatcher']]],
+  ['helpcommandhandler_60',['helpCommandHandler',['../classace__utils_1_1cli_1_1CommandDispatcher.html#af91ec1340b0b62e43dab869316cf9b22',1,'ace_utils::cli::CommandDispatcher']]],
+  ['helpspecific_61',['helpSpecific',['../classace__utils_1_1cli_1_1CommandDispatcher.html#aaba7e9cf696e4a84e0611cf9c9d40607',1,'ace_utils::cli::CommandDispatcher']]]
 ];