+++ docs/html/search/functions_8.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['streamlinereader_50',['StreamLineReader',['../classcli_1_1StreamLineReader.html#a1bb326fd74f0234f43083237f8e98ab7',1,'cli::StreamLineReader']]]
+  ['readwithcrc_67',['readWithCrc',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#a82c85ef76e886c832438a71d5d4e7e9b',1,'ace_utils::crc_eeprom::CrcEeprom']]],
+  ['run_68',['run',['../classace__utils_1_1cli_1_1CommandHandler.html#a82e846fdd080bfb36b13a1f903cb7def',1,'ace_utils::cli::CommandHandler']]],
+  ['runcommand_69',['runCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ae20ae336f3740b1faceecdf0babfa55e',1,'ace_utils::cli::CommandDispatcher']]]
 ];