+++ src/AceUtilsCli.h
@@ -1,0 +1,10 @@
+#ifndef ACE_UTILS_CLI_H
+#define ACE_UTILS_CLI_H
+
+#include "cli/InputLine.h"
+#include "cli/StreamLineReader.h"
+#include "cli/CommandDispatcher.h"
+#include "cli/CommandHandler.h"
+#include "cli/CommandManager.h"
+
+#endif